Very early implementation of property adjustment (velocity). Nowhere near right yet.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10504 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/gui/rulers/ControlItem.cpp
@@ -61,7 +61,7 @@ ControlItem::~ControlItem()
 //    delete m_elementAdapter;
 }
 
-void ControlItem::setValue(long v)
+void ControlItem::setValue(float v)
 {
 //     RG_DEBUG << "ControlItem::setValue(" << v << ") x = " << x() << endl;
